requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.29.0

pydantic==2.11.4
pydantic-settings==2.2.1

python-dotenv==1.0.1
transformers==4.40.0
langdetect==1.0.9
pymupdf==1.23.22
python-docx==1.1.0